Extract Logger::Stream into a separate file
